+++ sm-core/src/main/java/com/salesmanager/core/business/services/catalog/product/manufacturer/ManufacturerServiceImpl.java
@@ -1,17 +1,6 @@
 package com.salesmanager.core.business.services.catalog.product.manufacturer;
 
 
-import java.util.HashSet;
-import java.util.List;
-import javax.inject.Inject;
-import org.jsoup.helper.Validate;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
 import com.salesmanager.core.business.exception.ServiceException;
 import com.salesmanager.core.business.repositories.catalog.product.manufacturer.ManufacturerRepository;
 import com.salesmanager.core.business.repositories.catalog.product.manufacturer.PageableManufacturerRepository;
@@ -21,6 +10,17 @@ import com.salesmanager.core.model.catalog.product.manufacturer.Manufacturer;
 import com.salesmanager.core.model.catalog.product.manufacturer.ManufacturerDescription;
 import com.salesmanager.core.model.merchant.MerchantStore;
 import com.salesmanager.core.model.reference.language.Language;
+import org.jsoup.helper.Validate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+
+import javax.inject.Inject;
+import java.util.HashSet;
+import java.util.List;
 
 
 
@@ -90,7 +90,7 @@ public class ManufacturerServiceImpl extends SalesManagerEntityServiceImpl<Long,
 
     LOGGER.debug("Creating Manufacturer");
 
-    if (manufacturer.getId() != null && manufacturer.getId().longValue() > 0) {
+    if (manufacturer.getId() != null && manufacturer.getId() > 0) {
       super.update(manufacturer);
 
     } else {
